[SPOT-281][REFACTOR] 리팩토링 되지 않은 회원 애플리케이션 로직 분리 및 @Deprecated 애노테이션 추가
