+++ src/main/java/com/example/spot/common/config/WebSecurity.java
@@ -5,8 +5,8 @@ import com.example.spot.common.security.filters.JwtAuthenticationFilter;
 import com.example.spot.common.security.oauth.CustomOAuth2UserService;
 import com.example.spot.common.security.oauth.CustomOAuthSuccessHandler;
 import com.example.spot.common.security.utils.JwtTokenProvider;
-import com.example.spot.member.application.MemberService;
-import com.example.spot.member.application.UserDetailsServiceCustom;
+import com.example.spot.member.application.legacy.MemberService;
+import com.example.spot.member.application.legacy.UserDetailsServiceCustom;
 import lombok.RequiredArgsConstructor;
 
 import org.springframework.context.annotation.Bean;
