+++ src/main/java/com/example/spot/member/application/legacy/UserDetailsServiceCustom.java
@@ -1,6 +1,7 @@
-package com.example.spot.member.application;
+package com.example.spot.member.application.legacy;
 
 import org.springframework.security.core.userdetails.UserDetailsService;
 
+@Deprecated
 public interface UserDetailsServiceCustom extends UserDetailsService {
 }
