+++ src/main/java/com/example/spot/member/application/legacy/UserDetailsServiceCustomImpl.java
@@ -1,4 +1,4 @@
-package com.example.spot.member.application;
+package com.example.spot.member.application.legacy;
 
 import com.example.spot.member.domain.auth.TempUserDetails;
 import org.springframework.security.core.GrantedAuthority;
@@ -10,6 +10,7 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 @Service
+@Deprecated
 public class UserDetailsServiceCustomImpl implements UserDetailsServiceCustom {
 
     @Override
